pyproject.toml
[project]
name = "tiresias"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "brandon", email = "bmmaynard87@gmail.com" }
]
requires-python = ">=3.12.3"
dependencies = [
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyyaml>=6.0.3",
    "rich>=14.2.0",
    "typer>=0.21.1",
]

[project.scripts]
tiresias = "tiresias:main"

[build-system]
requires = ["uv_build>=0.9.24,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.11",
    "types-pyyaml>=6.0.12.20250915",
]
